refactor output table
